refactor to clean up api file
